Feat: created project's model
Created basic project set up with a test suite.
Created the app's model along with the DataModel class to handle the CRUD operations over the model. Also, I've created some functionality to build some test data on device's memory for testing purposes. This way we'll always have some values when running the app on the simulator as well as in the SwiftUI previews.

Closes #1