Feat: created tasks view
This is the view to show all the existing tasks so they could be edited, deleted or run.
The view also allows the user to search for a task by typing its name on the search bar and to order the tasks by name or creation date, sorting them in ascending or descending order.
Also, a plus button to create a new tasks is presented on the top right corner.

Closes #4