Feat: created steps view
The view to handle steps has been created to add or remove steps from a task, and to create new steps.
The functionality to remove a step has not been yet implemented.
Task and Tasks views have been updated to allow navigation to the steps view but none of them is finished.

Closes #17